refactored to be like sf

+++ src/engine/table.h
@@ -18,10 +18,10 @@ class table_entry
     uint64_t m_hash = 0;
     int32_t m_score = 0;
     chess::Move m_best_move = chess::Move::NULL_MOVE;
-    int16_t m_depth = 0;
+    int m_depth = 0;
     uint8_t m_flag = 0;
 
-    [[nodiscard]] table_entry_result get(uint64_t hash, int16_t ply, int16_t depth, int32_t alpha, int32_t beta) const
+    [[nodiscard]] table_entry_result get(uint64_t hash, int ply, int depth, int32_t alpha, int32_t beta) const
     {
         int32_t adj_score = 0;
         bool should_use = false;
@@ -63,7 +63,7 @@ class table_entry
         return {adj_score, should_use, best_move};
     }
 
-    void set(uint64_t hash, int32_t score, const chess::Move &best_move, int16_t ply, int16_t depth, uint8_t flag)
+    void set(uint64_t hash, int32_t score, const chess::Move &best_move, int ply, int depth, uint8_t flag)
     {
         m_hash = hash;
         m_depth = depth;
